Extract pure helpers out of the Player class
First step of the refactoring plan: move logic which does not need any
private state of the player into extension classes of the same
namespace, so all call sites stay unchanged, and move the three nested
classes into their own files.

Moved to extension classes:
- messages and localization (PlayerMessageExtensions)
- money and vault money (PlayerMoneyExtensions)
- item requirements, item destruction and invalid item logging
  (PlayerItemExtensions)
- self defense queries, next to the state they read (SelfDefenseExtensions)
- the magic effect power-up creation (MagicEffectPowerUpExtensions)
- the invisibility effect (PlayerInvisibilityExtensions)

Moved to their own files: PlayerAppearanceData (formerly the nested
AppearanceDataAdapter), GameMasterMagicEffectDefinition and
TemporaryItemStorage.

TryAddMoney loses its virtual modifier - no type overrides it.

No behavior change: Player.cs shrinks by 486 lines and the only
additions to it are the two renamed type references.
